Add modern event system flag + rename legacy plugin module
Add module
